Limit for hw decoder vertical resolution set to 1920. Internal improvements.
Vertical video resolutions of >~1900 pixels lead to incorrect vertical scaling
(seems to be a low-level hw or driver bug), thus, to avoid visible issues, the
limit has been set to 1920.

Tiny coding style improvements.
Updated nx/utils/ to match those of nx_vms.

<==== Mercurial data ====>
Branch: master

+++ artifacts/bpi/proxy-decoder/nx/utils/flag_config.cpp
@@ -60,7 +60,7 @@ static bool parseNameValue(const char* const s, std::string* outName, std::strin
         *outValue += *(p++);
 
     // Trim trailing spaces in value.
-    int i = outValue->size() - 1;
+    int i = (int) outValue->size() - 1;
     while (i >= 0 && isWhitespace((*outValue)[i]))
         --i;
     *outValue = outValue->substr(0, i + 1);
